+++ src/gui/qgsmaptoolcapture.h
@@ -38,6 +38,16 @@ class GUI_EXPORT QgsMapToolCapture : public QgsMapToolAdvancedDigitizing
     Q_OBJECT
 
   public:
+
+    //! Different capture modes
+    enum CaptureMode
+    {
+      CaptureNone,    //!< Do not capture / determine mode from layer geometry type
+      CapturePoint,   //!< Capture points
+      CaptureLine,    //!< Capture lines
+      CapturePolygon  //!< Capture polygons
+    };
+
     //! constructor
     QgsMapToolCapture( QgsMapCanvas *canvas, QgsAdvancedDigitizingDockWidget *cadDockWidget, CaptureMode mode );
 
@@ -45,6 +55,13 @@ class GUI_EXPORT QgsMapToolCapture : public QgsMapToolAdvancedDigitizing
 
     virtual void activate() override;
     virtual void deactivate() override;
+
+    /**
+     * The capture mode
+     *
+     * \returns Capture mode
+     */
+    CaptureMode mode() const { return mCaptureMode; }
 
     //! Adds a whole curve (e.g. circularstring) to the captured geometry. Curve must be in map CRS
     int addCurve( QgsCurve *c );
@@ -190,6 +207,9 @@ class GUI_EXPORT QgsMapToolCapture : public QgsMapToolAdvancedDigitizing
     bool tracingAddVertex( const QgsPointXY &point );
 
   private:
+    //! The capture mode in which this tool operates
+    CaptureMode mCaptureMode;
+
     //! Flag to indicate a map canvas capture operation is taking place
     bool mCapturing;
